Add logic to change square color is mouse "hovers" over it
Chose the 'mouseover' event because it fires when a pointer enters an
element or any of its children. This allows us to attach a since event
listener to the container and capture mouseover events for all of the
squares in the container. Adds an ID to each square the document can be
queried for the particular square. Then add a class to the square's
class list to add color to the square.