Make color darker the more times you pass over the square
Uses a CSS custom property to define the alpha for the background color.
While I have not learned about this in the Odin project yet, this is
the cleanest way to achieve this effect. When I tried using opacity, it
also made the square borders lighter. With this implementation, only the
background color changes while the borders stay the same.