Update how mouseover event identifies square
The event.target property holds a reference to the element and we could
check if that element's class list included the 'square' class instead
of checking if the ID started with 'square'. We also did not need to
then query the document for the particular square using its ID, again,
because the event contains a reference to that element.