Fix some formatting issues
Normalizes the use of single quotes vs double quotes. Now it's only
using double quotes inside prompts. Everywhere else is single quotes.

Removes unnecessary constant. CONTAINER_ID was originally being used in
the mouseover event. Since that use case was removed, the only other
place it was being used was selecting the container. Since it was being
used only in one place, I removed the constant.

Adds a semi-colon to the end of a couple of lines that were missing it.