Fix user input validation for setting squares per side
Original implementation did not allow the user to cancel, exit, or
refresh the page. Adds a check to see if the user has actually provided
an input. When the user clicks cancel or presses the escape key, the
user input will be falsy.

Update loop control to use a bool instead as recommend by one of the
videos on problem solving.